Normalize mobility values (#7)
Move mobilty common value/offset back to piece values, being able to better understand real mobility values and avoiding having weird piece values.

This is, at the end of the day, very similar to what is done with PSQTs. The main difference is that there I take the average of the table, and here I take the first non-zero value as offset.

Negative values at the beginning of bishop mobility array never get hit. That means that they're not excluded from the tuning, but they're never hit during it (for now)

https://github.com/lynx-chess/Lynx/pull/788